+++ web3.js/src/message.ts
@@ -5,7 +5,7 @@ import * as BufferLayout from '@solana/buffer-layout';
 import {PublicKey} from './publickey';
 import type {Blockhash} from './blockhash';
 import * as Layout from './layout';
-import {PACKET_DATA_SIZE} from './transaction-constants';
+import {PACKET_DATA_SIZE} from './transaction/constants';
 import * as shortvec from './util/shortvec-encoding';
 import {toBuffer} from './util/to-buffer';
