+++ web3.js/src/transaction/index.ts
@@ -1,0 +1,2 @@
+export * from './constants';
+export * from './legacy';
