+++ web3.js/src/transaction/legacy.ts
@@ -2,19 +2,16 @@ import nacl from 'tweetnacl';
 import bs58 from 'bs58';
 import {Buffer} from 'buffer';
 
-import {
-  PACKET_DATA_SIZE,
-  SIGNATURE_LENGTH_IN_BYTES,
-} from './transaction-constants';
-import {Connection} from './connection';
-import {Message} from './message';
-import {PublicKey} from './publickey';
-import * as shortvec from './util/shortvec-encoding';
-import {toBuffer} from './util/to-buffer';
-import invariant from './util/assert';
-import type {Signer} from './keypair';
-import type {Blockhash} from './blockhash';
-import type {CompiledInstruction} from './message';
+import {PACKET_DATA_SIZE, SIGNATURE_LENGTH_IN_BYTES} from './constants';
+import {Connection} from '../connection';
+import {Message} from '../message';
+import {PublicKey} from '../publickey';
+import * as shortvec from '../util/shortvec-encoding';
+import {toBuffer} from '../util/to-buffer';
+import invariant from '../util/assert';
+import type {Signer} from '../keypair';
+import type {Blockhash} from '../blockhash';
+import type {CompiledInstruction} from '../message';
 
 /**
  * Transaction signature as base-58 encoded string
